requirements.txt
clint== 0.5.1
pyfiglet== 0.8.post1
gitpython== 3.1.0
jira== 2.0.0
